Added backtesting
Added a bunch of changes to Trading_Strategy.py that should work alongside backtest_engine.py which will backtest a random strategy for a stock using JSON and Gemini API.

I didn't test out if manual entry works, since I have little trading knowledge, but an easy way to test is just to ask ChatGPT or another LLM to give you a trading strategy in either the JSON format or the Python format. Once you do that you should be able to test it out to see if it works.

Otherwise, this is a basic and starting point for backtesting trading strategies. Can see some improvement later.